implement reading alf files with columns description in metadata
