+++ tests/ibllib/test_io.py
@@ -3,12 +3,10 @@ import os
 import uuid
 import tempfile
 from pathlib import Path
-import shutil
 
 import numpy as np
 
 from ibllib.io import params, flags, jsonable, spikeglx
-import alf.io
 
 
 class TestsParams(unittest.TestCase):
@@ -175,80 +173,6 @@ class TestsSpikeGLX(unittest.TestCase):
         for m in range(1, 16):
             self.assertEqual(np.sum(out[m]), 1)
             self.assertEqual(out[m, m - 1], 1)
-
-
-class TestsAlf(unittest.TestCase):
-    def setUp(self) -> None:
-        self.tmpdir = Path(tempfile.gettempdir()) / 'iotest'
-        self.tmpdir.mkdir(exist_ok=True)
-        self.vfile = self.tmpdir / 'toto.titi.npy'
-        self.tfile = self.tmpdir / 'toto.timestamps.npy'
-        self.object_files = [self.tmpdir / 'neuveu.riri.npy',
-                             self.tmpdir / 'neuveu.fifi.npy',
-                             self.tmpdir / 'neuveu.loulou.npy']
-        for f in self.object_files:
-            np.save(file=f, arr=np.random.rand(5,))
-
-    def test_read_ts(self):
-        # simplest test possible with one column in each file
-        t = np.arange(0, 10)
-        d = np.random.rand(10)
-        np.save(self.vfile, d)
-        np.save(self.tfile, t)
-        t_, d_ = alf.io.read_ts(self.vfile)
-        self.assertTrue(np.all(t_ == t))
-        self.assertTrue(np.all(d_ == d))
-
-    def test_load_object(self):
-        # first usage of load object is to provide one of the files belonging to the object
-        obj = alf.io.load_object(self.object_files[0])
-        self.assertTrue(set(obj.keys()) == set(['riri', 'fifi', 'loulou']))
-        self.assertTrue(all([obj[o].shape == (5,) for o in obj]))
-        # the second usage is to provide a directory and the object name
-        obj = alf.io.load_object(self.tmpdir, 'neuveu')
-        self.assertTrue(set(obj.keys()) == set(['riri', 'fifi', 'loulou']))
-        self.assertTrue(all([obj[o].shape == (5,) for o in obj]))
-        # and this should throw a value error
-        with self.assertRaises(ValueError) as context:
-            obj = alf.io.load_object(self.tmpdir)
-        self.assertTrue('object name should be provided too' in str(context.exception))
-
-    def test_save_npy(self):
-        # test with straight vectors
-        a = {'riri': np.random.rand(100),
-             'fifi': np.random.rand(100)}
-        alf.io.save_object_npy(self.tmpdir, a, 'neuveux')
-        # read after write
-        b = alf.io.load_object(self.tmpdir, 'neuveux')
-        for k in a:
-            self.assertTrue(np.all(a[k] == b[k]))
-        # test with more exotic shapes, still valid
-        a = {'riri': np.random.rand(100),
-             'fifi': np.random.rand(100, 2),
-             'loulou': np.random.rand(1, 2)}
-        alf.io.save_object_npy(self.tmpdir, a, 'neuveux')
-        # read after write
-        b = alf.io.load_object(self.tmpdir, 'neuveux')
-        for k in a:
-            self.assertTrue(np.all(a[k] == b[k]))
-        # test with non allowed shape
-        a = {'riri': np.random.rand(100),
-             'fifi': np.random.rand(100, 2),
-             'loulou': np.random.rand(5, 2)}
-        with self.assertRaises(Exception) as context:
-            alf.io.save_object_npy(self.tmpdir, a, 'neuveux')
-        self.assertTrue('Dimensions are not consistent' in str(context.exception))
-
-    def test_check_dimensions(self):
-        a = {'a': np.ones([10, 10]), 'b': np.ones([10, 2]), 'c': np.ones([10])}
-        status = alf.io.check_dimensions(a)
-        self.assertTrue(status == 1)
-        a = {'a': np.ones([10, 10]), 'b': np.ones([10, 1]), 'c': np.ones([10])}
-        status = alf.io.check_dimensions(a)
-        self.assertTrue(status == 0)
-
-    def tearDown(self) -> None:
-        shutil.rmtree(self.tmpdir)
 
 
 if __name__ == "__main__":
